+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+- Add support for self-hosted instances
+- Add support for non-secure requests (HTTP)
+
 ## 0.1.1
 - Bug fixes